split some classes and enumerations in sandbox projects
