+++ sandbox/contraints.hpp
@@ -1,0 +1,46 @@
+#ifndef MATADOR_CONSTRAINTS_HPP
+#define MATADOR_CONSTRAINTS_HPP
+
+#include <unordered_map>
+
+namespace matador {
+
+enum class constraints : unsigned char {
+  NONE =        0,
+  NOT_NULL =    1 << 0,
+  INDEX =       1 << 1,
+  UNIQUE =      1 << 2,
+  PRIMARY_KEY = 1 << 3,
+  FOREIGN_KEY = 1 << 4,
+  DEFAULT =     1 << 5,
+  UNIQUE_NOT_NULL = UNIQUE | NOT_NULL
+};
+
+std::unordered_map<constraints, std::string> constraints_to_name_map {
+{ constraints::NONE, "none" },
+{ constraints::NOT_NULL, "not null" },
+{ constraints::INDEX, "index" },
+{ constraints::UNIQUE, "unique" },
+{ constraints::PRIMARY_KEY, "primary key" },
+{ constraints::FOREIGN_KEY, "foreign key" },
+{ constraints::DEFAULT, "default" }
+};
+
+inline constraints operator|(constraints a, constraints b)
+{
+  return static_cast<constraints>(static_cast<unsigned int>(a) | static_cast<unsigned int>(b));
+}
+
+inline constraints operator&(constraints a, constraints b)
+{
+  return static_cast<constraints>(static_cast<unsigned int>(a) & static_cast<unsigned int>(b));
+}
+
+inline bool is_constraint_set(constraints source, constraints needle)
+{
+  return static_cast<int>(source & needle) > 0;
+}
+
+}
+
+#endif //MATADOR_CONSTRAINTS_HPP
